Move utility functions to new file
